+++ euporie/notebook/_commands.py
@@ -44,3 +44,11 @@ def _view_documentation() -> None:
     from euporie.core.app.current import get_app
 
     get_app().open_file(UPath("https://euporie.readthedocs.io/"))
+
+
+@add_cmd()
+def _toggle_side_bar_pane() -> None:
+    """Open or close the current side-bar pane."""
+    from euporie.notebook.current import get_app
+
+    get_app().side_bar.toggle_pane()